# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/clerk/goods": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Get CartList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "oauth token",
                        "name": "State",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "oauth token",
                        "name": "oauthstate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CartLists.GoodInCart"
                        }
                    },
                    "400": {
                        "description": "invalid_parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "fail get goods in cart",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clerk"
                ],
                "summary": "Add Goods",
                "parameters": [
                    {
                        "type": "string",
                        "description": "image name",
                        "name": "ImageName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "descript",
                        "name": "Descript",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price",
                        "name": "Price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "class",
                        "name": "Class",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "invalid_insert_parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/clerk/goods/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clerk"
                ],
                "summary": "Clerk Get Goods",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid_parameter",
                        "schema": {
                            "type": "page"
                        }
                    },
                    "404": {
                        "description": "no_found_item",
                        "schema": {
                            "type": "page"
                        }
                    }
                }
            }
        },
        "/api/v1/clerk/goods/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clerk"
                ],
                "summary": "Update Goods",
                "parameters": [
                    {
                        "type": "string",
                        "description": "image name",
                        "name": "ImageName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "descript",
                        "name": "Descript",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price",
                        "name": "Price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "class",
                        "name": "Class",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "invalid update parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clerk"
                ],
                "summary": "Update Goods",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "invalid delete parameter",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/customer/goods": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Add Cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "oauth token",
                        "name": "State",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "oauth token",
                        "name": "oauthstate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Good param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddCartGoods.GoodsParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid_parameter",
                        "schema": {
                            "type": "page"
                        }
                    },
                    "404": {
                        "description": "no_found_item",
                        "schema": {
                            "type": "page"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Delete Cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "oauth token",
                        "name": "State",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "oauth token",
                        "name": "oauthstate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Good param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteCartGoods.GoodsParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid_parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "fail get goods in cart",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/customer/order": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Creat Order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "oauth token",
                        "name": "State",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "oauth token",
                        "name": "oauthstate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "cart param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateOrder.GoodsParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "fail create order",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.AddCartGoods.GoodsParams": {
            "type": "object",
            "required": [
                "email",
                "good_amount",
                "good_name",
                "good_price"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "good_amount": {
                    "type": "integer"
                },
                "good_name": {
                    "type": "string"
                },
                "good_price": {
                    "type": "integer"
                }
            }
        },
        "v1.CartLists.GoodInCart": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "image_name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "v1.CreateOrder.GoodsParams": {
            "type": "object",
            "required": [
                "email",
                "good_amount",
                "good_name",
                "message"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "good_amount": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "good_name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "v1.DeleteCartGoods.GoodsParams": {
            "type": "object",
            "required": [
                "email",
                "good_amount",
                "good_name",
                "good_price"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "good_amount": {
                    "type": "integer"
                },
                "good_name": {
                    "type": "string"
                },
                "good_price": {
                    "type": "integer"
                }
            }
        }
    }
}